Bug fix for issue #5.
Moved calculation of max_blocks to countpairs_s_mu_mocks_impl (and countpairs_s_mu_impl).
Thus the block LUTs will be allocated with the correct size.  Also increased default
max_threads from 400M to 1.6B.  And changed num_integer_weights type from uint_8 to int.

+++ mocks/DDsmu_mocks/countpairs_s_mu_mocks_gpu.cu
@@ -40,7 +40,7 @@ typedef struct
     //pair_weight_struct_double pair_weight;
 
     int num_weights;
-    uint8_t num_integer_weights;
+    int num_integer_weights;
     int8_t noffset;
     double default_value;
 } pair_struct_double;
@@ -63,7 +63,7 @@ typedef struct
     //pair_weight_struct_float pair_weight;
 
     int num_weights;
-    uint8_t num_integer_weights;
+    int num_integer_weights;
     int8_t noffset;
     float default_value;
 } pair_struct_float;
